requirements.txt
Django==2.1.7
django-ckeditor==5.6.1
pillow==5.4.1
python-decouple==3.1
djangorestframework==3.9.2
djongo==1.2.31 
dnspython==1.16.0
